Show notice when preferred positioning method change doesn't stick
In some situations the PositionSource doesn't accept setting the
preferred positionin methods to some value. Instead it changes it by
itself to some other value. If this happens, show a notice to the user.

The preferred positioning methods combobox now also always reflects the
actual value so it doesn't have to be shown in the label anymore.